+++ packages/vfs/CHANGELOG.md
@@ -1,5 +1,67 @@
 # @tinycloud/vfs
 
+## 0.1.9-beta.1
+
+### Patch Changes
+
+- Updated dependencies [367c17c]
+  - @tinycloud/node-sdk@2.6.4-beta.1
+
+## 0.1.9-beta.0
+
+### Patch Changes
+
+- Updated dependencies [f6048b7]
+  - @tinycloud/node-sdk@2.6.4-beta.0
+
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies [3841be4]
+  - @tinycloud/node-sdk@2.6.3
+
+## 0.1.8-beta.0
+
+### Patch Changes
+
+- Updated dependencies [3841be4]
+  - @tinycloud/node-sdk@2.6.3-beta.0
+
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [b4d1e45]
+  - @tinycloud/node-sdk@2.6.2
+
+## 0.1.7-beta.0
+
+### Patch Changes
+
+- Updated dependencies [b4d1e45]
+  - @tinycloud/node-sdk@2.6.2-beta.0
+
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [bf31506]
+  - @tinycloud/node-sdk@2.6.1
+
+## 0.1.6-beta.1
+
+### Patch Changes
+
+- @tinycloud/node-sdk@2.6.1-beta.1
+
+## 0.1.6-beta.0
+
+### Patch Changes
+
+- Updated dependencies [bf31506]
+  - @tinycloud/node-sdk@2.6.1-beta.0
+
 ## 0.1.5
 
 ### Patch Changes