+++ tests/sdk-services/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @tinycloudlabs/sdk-services-test
 
+## 11.0.3
+
+### Patch Changes
+
+- Updated dependencies [3841be4]
+  - @tinycloud/sdk-services@2.6.3
+
+## 11.0.3-beta.0
+
+### Patch Changes
+
+- Updated dependencies [3841be4]
+  - @tinycloud/sdk-services@2.6.3-beta.0
+
+## 11.0.2
+
+### Patch Changes
+
+- Updated dependencies [b4d1e45]
+  - @tinycloud/sdk-services@2.6.2
+
+## 11.0.2-beta.0
+
+### Patch Changes
+
+- Updated dependencies [b4d1e45]
+  - @tinycloud/sdk-services@2.6.2-beta.0
+
+## 11.0.1
+
+### Patch Changes
+
+- Updated dependencies [bf31506]
+  - @tinycloud/sdk-services@2.6.1
+
+## 11.0.1-beta.1
+
+### Patch Changes
+
+- @tinycloud/sdk-services@2.6.1-beta.1
+
+## 11.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [bf31506]
+  - @tinycloud/sdk-services@2.6.1-beta.0
+
 ## 11.0.0
 
 ### Patch Changes